proyecto/requirements.txt
Django==4.1.1
selenium==4.4.3
webdriver-manager==3.8.3
pandas==1.4.4
requests==2.28.1
djangorestframework==3.13.1
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cron==0.6.0
psycopg2==2.9.3